+++ packages/opentelemetry-instrumentation-openai/tests/traces/test_completions.py
@@ -1,3 +1,5 @@
+import os
+
 import pytest
 
 
@@ -65,28 +67,46 @@ def test_completion_langchain_style(exporter, openai_client):
 
 @pytest.mark.vcr
 def test_completion_streaming(exporter, openai_client):
-    response = openai_client.completions.create(
-        model="davinci-002",
-        prompt="Tell me a joke about opentelemetry",
-        stream=True,
-    )
+    # set os env for token usage record in stream mode
+    original_value = os.environ.get("TRACELOOP_STREAM_TOKEN_USAGE")
+    os.environ["TRACELOOP_STREAM_TOKEN_USAGE"] = "true"
 
-    for _ in response:
-        pass
+    try:
+        response = openai_client.completions.create(
+            model="davinci-002",
+            prompt="Tell me a joke about opentelemetry",
+            stream=True,
+        )
 
-    spans = exporter.get_finished_spans()
-    assert [span.name for span in spans] == [
-        "openai.completion",
-    ]
-    open_ai_span = spans[0]
-    assert (
-        open_ai_span.attributes["llm.prompts.0.user"]
-        == "Tell me a joke about opentelemetry"
-    )
-    assert open_ai_span.attributes.get("llm.completions.0.content")
-    assert (
-        open_ai_span.attributes.get("openai.api_base") == "https://api.openai.com/v1/"
-    )
+        for _ in response:
+            pass
+
+        spans = exporter.get_finished_spans()
+        assert [span.name for span in spans] == [
+            "openai.completion",
+        ]
+        open_ai_span = spans[0]
+        assert (
+            open_ai_span.attributes["llm.prompts.0.user"]
+            == "Tell me a joke about opentelemetry"
+        )
+        assert open_ai_span.attributes.get("llm.completions.0.content")
+        assert (
+            open_ai_span.attributes.get("openai.api_base") == "https://api.openai.com/v1/"
+        )
+
+        # check token usage attributes for stream
+        completion_tokens = open_ai_span.attributes.get("llm.usage.completion_tokens")
+        prompt_tokens = open_ai_span.attributes.get("llm.usage.prompt_tokens")
+        total_tokens = open_ai_span.attributes.get("llm.usage.total_tokens")
+        assert completion_tokens and prompt_tokens and total_tokens
+        assert completion_tokens + prompt_tokens == total_tokens
+    finally:
+        # unset env
+        if original_value is None:
+            del os.environ["TRACELOOP_STREAM_TOKEN_USAGE"]
+        else:
+            os.environ["TRACELOOP_STREAM_TOKEN_USAGE"] = original_value
 
 
 @pytest.mark.vcr