+++ packages/traceloop-sdk/traceloop/sdk/__init__.py
@@ -167,7 +167,8 @@ class Traceloop:
 
         # Metrics init: disabled for Traceloop as we don't have a metrics endpoint (yet)
         if api_endpoint.find("traceloop.com") != -1 or not is_metrics_enabled():
-            print(Fore.YELLOW + "Metrics is disabled" + Fore.RESET)
+            if not is_metrics_enabled():
+                print(Fore.YELLOW + "OpenTelemetry metrics are disabled" + Fore.RESET)
             return
 
         if metrics_exporter: