+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## v0.14.5 (2024-03-21)
+
+### Fix
+
+- **anthropic**: support messages API (#671)
+
+## v0.14.4 (2024-03-21)
+
+### Fix
+
+- auto-instrumentation support (#662)
+- **sample**: poetry issues; litellm sample
+- **sdk**: better logging for otel metrics
+- **sdk**: error for manually providing instrumentation list
+
+## v0.14.3 (2024-03-17)
+
+### Fix
+
+- support python 3.12 (#639)
+- **traceloop-sdk**: Log error message when providing wrong API key. (#638)
+
+## v0.14.2 (2024-03-15)
+
+### Fix
+
+- **openai**: support tool syntax (#630)
+
 ## v0.14.1 (2024-03-12)
 
 ### Fix